[commands,messaging,statistics] fix non-deleted messages deadlock
Add error to individual messages so that non-decryptable messages can
simply be removed rather than causing the system to deadlock.

The original code that error on the first failing message would loose
all good messages, this version will allow all good messages to be
processed even if failing messages are mixed in.

Signed-off-by: Christopher Hall <hsw@bitmark.com>